imports and remove scratch code
